Get rid of chrono alltogether
`chrono` is conflictive with embedded scenarios, as relies on `libc`
functions that are not supported on all architectures, even if libstd
is available for those architectures.

This commit takes the assumption that the timestamp present in the
type `TangleMessage` is a leftover from pre-chrysalis. It's the only
place where `chrono` is needed. Therefore, the following actions have
been taken:

- Convert `TangleMessage<F>` into just a type alias of `BinaryMessage<F, TangleAddress>`
- Remove `chrono` from `streams-app` dependencies