Specialice `random_key` for espidf OS targets
espidf targets have libstd available through ESP-IDF framework.
This framework provides most but not all of the POSIX standard API.
In particular, ESP-IDF does not provide a means to fork threads, and
cannot implement `pthread_atfork` (see
https://github.com/espressif/rust-esp32-example/issues/23).
This means `rand::thread_rng()` is not available on those targets,
as it is based on storing the rng in the thread-local storage, accessed
through a thread handle.

For these targets, this commit specializes the random_key and
random_nonce generators to use `StdRng::from_entropy()` instead.
In practice, given the coldness of these 2 functions, the same level
of security and performance is to be expected.

Note that currently `getrandom` does not support `espidf` targets
and requires a small patch.