Remove `PhantomData<F>`from `BinaryBody`
Some of the `PhantomData<F>` used throughtout the project are debatable.
They feel like necessary because type arguments
and bounds are placed too early in `impl` blocks instead of the very same
methods that need them. However, type arguments in data types produce a
cascade effect that infects everything the type passes by, introducing a significant
amount of complexity, unreadability and, in some cases, actual contention to achieve certain things.

One of the more evident cases is `BinaryBody`, which from a conceptual
point of view shouldn't ever concern with the `PRP` except for when
being wrapped or unwrapped, as we already have `WrappedMessage` and
`UnwrappedMessage` that act as bundles of a message and a spongos state.