Fix WASM and C examples after `User::reset_state()` behaviour  change
Commit 655d28c fixed `User::reset_state()` by resetting also the
link-store of the user. This is considered the proper behaviour of
`User::reset_state()`, as a the common expectancy of a reseted user is
for it to be as blank as possible, keeping only its identity data.

This change, makes resetted users lose any data of the channel
whatsoever, including subscription messages that cannot be recovered
by fetching from the Tangle. This makes it a method only to be used
in specific scenarios where for security reasons a user needs to trully
be reseted; it also renders it not suitable as a means to _just_ replay
the channel messages from the announcement message.

The examples fixed in this commit were exemplifying the unsubscribe
feature after `User::reset_state()` was called, and broke after the
explained change for the aforementioned reason.